bookmark app week9 changes

+++ bookmark/database/migrations/2022_03_26_045539_create_books_table.php
@@ -20,7 +20,7 @@ class CreateBooksTable extends Migration
             $table->string('slug');
             $table->string('title');
             $table->string('author');
-            $table->tinyInteger('published_year');
+            $table->smallInteger('published_year');
             #  $table->string('cover_url')->nullable(); # Here we use a column modifier to specify this column can be left empty (null)
             $table->string('cover_url');
             $table->string('info_url');
